jool: split into several meta-packages
This commit is about saving space in different setups.
It is hard to fit Jool on a 4MB device. For basic configuration the userspace tools are not needed. 
Furthermore Jool offers two different things - a NAT64 and a NAT46.
This commit creates these new packages:
* kmod-jool-stateful
* kmod-jool-stateless
* jool-tools
** jool-tools-stateful 
** jool-tools-stateless

Existing packages are being converted to meta-packages with dependencies to the corresponding new packages:
* jool selects kmod-jool and jool-tools
* kmod-jool selects kmod-jool-stateful and kmod-jool-stateless
* jool-tools selects jool-tools-stateful and jool-tools-stateless